Fix Oracle Last-Modified headers
Oracle's SWIFT implementation returns Last-Modified headers with "UTC"
timezone, which will cause an error when trying to parse the time with
net/http.TimeFormat because the TZ should be "GMT" instead of "UTC".

Since the parsing is done within the `swift` package we depend on, we
are fixing the headers with a custom transport until we can raise the
issue with the maintainer of our dependency.