Local: mimic the path semantics of cloud backends (#175)
* Local: mimic the path semantics of cloud backends

When accessing a container based on its location,
or an item based on its container, interpret relative paths
as relative to the location (in the case of container) or the
container (in the case of an item).

This is analogous to cloud backends, and so container.Item("name")
means the same thing in all backends - it means an item called "name"
that is stored inside the container.

On the other hand, we still want to support
container.Item("/full/path/to/name"), because the full path is the ID
of the Item. So this patch interprets an absolute path as the full OS
path of the item.

* Add tests for fetching container and item by name

Add backend-agnostic tests for the following scenarios:
- Given location and name of container, open the container.
- Given container and path inside container, open item

* LOCAL: add a specific test for ItemByURL

Add a test that imposes the convention that in an item created
from a URL, the item's name is the base file name.