local: only walk the subdirectory designated by prefix
When listing items using the local-file-system driver, only
traverse the subdirectory with longest prefix that contains
the given listing prefix.

Three cases are of interest:

a. The listing prefix is foo/bar/ - we walk subdirectory foo/bar
   under the container's path.
b. The listing prefix is foo/ba - we walk subdirectory foo under
   the container's path, and filter stuff that doesn't begin with
   foo/ba.
c. The prefix is an empty string - walk the container's path like
   today.