Local: allow -1 as unknown size on Put
When putting an item, the item's data is retrieved from an
io.Reader, but a size argument is also supplied. This change
causes the local backend to ignore the size (as what most other
backends do) if it's negative. This allows for data sources whose
size is not known in advance (e.g. getting an HTTP upload request
with a multipart/form-data payload).